Add style_traits crate to improve crate separation, closes #7353
